% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/paws.sms_operations.R
\name{create_replication_job}
\alias{create_replication_job}
\title{The CreateReplicationJob API is used to create a ReplicationJob to replicate a server on AWS}
\usage{
create_replication_job(serverId, seedReplicationTime, frequency,
  licenseType = NULL, roleName = NULL, description = NULL)
}
\arguments{
\item{serverId}{[required]}

\item{seedReplicationTime}{[required]}

\item{frequency}{[required]}

\item{licenseType}{}

\item{roleName}{}

\item{description}{}
}
\description{
The CreateReplicationJob API is used to create a ReplicationJob to replicate a server on AWS. Call this API to first create a ReplicationJob, which will then schedule periodic ReplicationRuns to replicate your server to AWS. Each ReplicationRun will result in the creation of an AWS AMI.
}
\section{Accepted Parameters}{
\preformatted{create_replication_job(
  serverId = "string",
  seedReplicationTime = as.POSIXct("2015-01-01"),
  frequency = 123,
  licenseType = "AWS"|"BYOL",
  roleName = "string",
  description = "string"
)
}
}

