% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/paws.sms_operations.R
\name{delete_app_replication_configuration}
\alias{delete_app_replication_configuration}
\title{Deletes existing replication configuration for an application}
\usage{
delete_app_replication_configuration(
  appId = "string"
)
}
\arguments{
\item{appId}{ID of the application associated with the replication configuration.}
}
\description{
Deletes existing replication configuration for an application.
}
\examples{

}
