% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/paws.sms_operations.R
\name{delete_replication_job}
\alias{delete_replication_job}
\title{The DeleteReplicationJob API is used to delete a ReplicationJob, resulting in no further ReplicationRuns}
\usage{
delete_replication_job(replicationJobId)
}
\arguments{
\item{replicationJobId}{[required]}
}
\description{
The DeleteReplicationJob API is used to delete a ReplicationJob, resulting in no further ReplicationRuns. This will delete the contents of the S3 bucket used to store SMS artifacts, but will not delete any AMIs created by the SMS service.
}
\section{Accepted Parameters}{
\preformatted{delete_replication_job(
  replicationJobId = "string"
)
}
}

