% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/paws.sms_operations.R
\name{delete_server_catalog}
\alias{delete_server_catalog}
\title{Deletes all servers from your server catalog}
\usage{
delete_server_catalog()
}
\description{
Deletes all servers from your server catalog.
}
\examples{

}
