% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/paws.sms_operations.R
\name{generate_template}
\alias{generate_template}
\title{Generates an Amazon CloudFormation template based on the current launch configuration and writes it to an Amazon S3 object in the customer's Amazon S3 bucket}
\usage{
generate_template(
  appId = "string",
  templateFormat = "JSON"|"YAML"
)
}
\arguments{
\item{appId}{ID of the application associated with the Amazon CloudFormation template.}

\item{templateFormat}{Format for generating the Amazon CloudFormation template.}
}
\description{
Generates an Amazon CloudFormation template based on the current launch configuration and writes it to an Amazon S3 object in the customer's Amazon S3 bucket.
}
\examples{

}
