% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/paws.sms_operations.R
\name{get_app}
\alias{get_app}
\title{Retrieve information about an application}
\usage{
get_app(
  appId = "string"
)
}
\arguments{
\item{appId}{ID of the application whose information is being retrieved.}
}
\description{
Retrieve information about an application.
}
\examples{

}
