% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/paws.sms_operations.R
\name{get_app_launch_configuration}
\alias{get_app_launch_configuration}
\title{Retrieves the application launch configuration associated with an application}
\usage{
get_app_launch_configuration(appId = NULL)
}
\arguments{
\item{appId}{ID of the application launch configuration.}
}
\description{
Retrieves the application launch configuration associated with an application.
}
\section{Accepted Parameters}{
\preformatted{get_app_launch_configuration(
  appId = "string"
)
}
}

