% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/paws.sms_operations.R
\name{get_app_replication_configuration}
\alias{get_app_replication_configuration}
\title{Retrieves an application replication configuration associatd with an application}
\usage{
get_app_replication_configuration(appId = NULL)
}
\arguments{
\item{appId}{ID of the application associated with the replication configuration.}
}
\description{
Retrieves an application replication configuration associatd with an application.
}
\section{Accepted Parameters}{
\preformatted{get_app_replication_configuration(
  appId = "string"
)
}
}

