% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/paws.sms_operations.R
\name{get_connectors}
\alias{get_connectors}
\title{The GetConnectors API returns a list of connectors that are registered with the Server Migration Service}
\usage{
get_connectors(nextToken = NULL, maxResults = NULL)
}
\arguments{
\item{nextToken}{}

\item{maxResults}{}
}
\description{
The GetConnectors API returns a list of connectors that are registered with the Server Migration Service.
}
\section{Accepted Parameters}{
\preformatted{get_connectors(
  nextToken = "string",
  maxResults = 123
)
}
}

