% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/paws.sms_operations.R
\name{get_replication_jobs}
\alias{get_replication_jobs}
\title{Describes the specified replication job or all of your replication jobs}
\usage{
get_replication_jobs(
  replicationJobId = "string",
  nextToken = "string",
  maxResults = 123
)
}
\arguments{
\item{replicationJobId}{The identifier of the replication job.}

\item{nextToken}{The token for the next set of results.}

\item{maxResults}{The maximum number of results to return in a single call. The default value is 50. To retrieve the remaining results, make another call with the returned \code{NextToken} value.}
}
\description{
Describes the specified replication job or all of your replication jobs.
}
\examples{

}
