% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/paws.sms_operations.R
\name{get_replication_runs}
\alias{get_replication_runs}
\title{Describes the replication runs for the specified replication job}
\usage{
get_replication_runs(
  replicationJobId = "string",
  nextToken = "string",
  maxResults = 123
)
}
\arguments{
\item{replicationJobId}{[required] The identifier of the replication job.}

\item{nextToken}{The token for the next set of results.}

\item{maxResults}{The maximum number of results to return in a single call. The default value is 50. To retrieve the remaining results, make another call with the returned \code{NextToken} value.}
}
\description{
Describes the replication runs for the specified replication job.
}
\examples{

}
