% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/paws.sms_operations.R
\name{get_servers}
\alias{get_servers}
\title{Describes the servers in your server catalog}
\usage{
get_servers(nextToken = NULL, maxResults = NULL,
  vmServerAddressList = NULL)
}
\arguments{
\item{nextToken}{The token for the next set of results.}

\item{maxResults}{The maximum number of results to return in a single call. The default value is 50. To retrieve the remaining results, make another call with the returned \code{NextToken} value.}

\item{vmServerAddressList}{List of \code{VmServerAddress} objects}
}
\description{
Describes the servers in your server catalog.
}
\details{
Before you can describe your servers, you must import them using ImportServerCatalog.
}
\section{Accepted Parameters}{
\preformatted{get_servers(
  nextToken = "string",
  maxResults = 123,
  vmServerAddressList = list(
    list(
      vmManagerId = "string",
      vmId = "string"
    )
  )
)
}
}

