% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/paws.sms_operations.R
\name{import_server_catalog}
\alias{import_server_catalog}
\title{Gathers a complete list of on-premises servers}
\usage{
import_server_catalog()
}
\description{
Gathers a complete list of on-premises servers. Connectors must be installed and monitoring all servers that you want to import.
}
\details{
This call returns immediately, but might take additional time to retrieve all the servers.
}
\examples{

}
