% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/paws.sms_operations.R
\name{launch_app}
\alias{launch_app}
\title{Launches an application stack}
\usage{
launch_app(appId = NULL)
}
\arguments{
\item{appId}{ID of the application to launch.}
}
\description{
Launches an application stack.
}
\section{Accepted Parameters}{
\preformatted{launch_app(
  appId = "string"
)
}
}

