% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/paws.sms_operations.R
\name{list_apps}
\alias{list_apps}
\title{Returns a list of summaries for all applications}
\usage{
list_apps(appIds = NULL, nextToken = NULL, maxResults = NULL)
}
\arguments{
\item{appIds}{}

\item{nextToken}{The token for the next set of results.}

\item{maxResults}{The maximum number of results to return in a single call. The default value is 50. To retrieve the remaining results, make another call with the returned \code{NextToken} value.}
}
\description{
Returns a list of summaries for all applications.
}
\section{Accepted Parameters}{
\preformatted{list_apps(
  appIds = list(
    "string"
  ),
  nextToken = "string",
  maxResults = 123
)
}
}

