% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/paws.sms_operations.R
\name{put_app_launch_configuration}
\alias{put_app_launch_configuration}
\title{Creates a launch configuration for an application}
\usage{
put_app_launch_configuration(
  appId = "string",
  roleName = "string",
  serverGroupLaunchConfigurations = list(
    list(
      serverGroupId = "string",
      launchOrder = 123,
      serverLaunchConfigurations = list(
        list(
          server = list(
            serverId = "string",
            serverType = "VIRTUAL_MACHINE",
            vmServer = list(
              vmServerAddress = list(
                vmManagerId = "string",
                vmId = "string"
              ),
              vmName = "string",
              vmManagerName = "string",
              vmManagerType = "VSPHERE"|"SCVMM"|"HYPERV-MANAGER",
              vmPath = "string"
            ),
            replicationJobId = "string",
            replicationJobTerminated = TRUE|FALSE
          ),
          logicalId = "string",
          vpc = "string",
          subnet = "string",
          securityGroup = "string",
          ec2KeyName = "string",
          userData = list(
            s3Location = list(
              bucket = "string",
              key = "string"
            )
          ),
          instanceType = "string",
          associatePublicIpAddress = TRUE|FALSE
        )
      )
    )
  )
)
}
\arguments{
\item{appId}{ID of the application associated with the launch configuration.}

\item{roleName}{Name of service role in the customer's account that Amazon CloudFormation uses to launch the application.}

\item{serverGroupLaunchConfigurations}{Launch configurations for server groups in the application.}
}
\description{
Creates a launch configuration for an application.
}
\examples{

}
