% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/paws.sms_operations.R
\name{start_app_replication}
\alias{start_app_replication}
\title{Starts replicating an application}
\usage{
start_app_replication(
  appId = "string"
)
}
\arguments{
\item{appId}{ID of the application to replicate.}
}
\description{
Starts replicating an application.
}
\examples{

}
