% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/paws.sms_operations.R
\name{terminate_app}
\alias{terminate_app}
\title{Terminates the stack for an application}
\usage{
terminate_app(appId = NULL)
}
\arguments{
\item{appId}{ID of the application to terminate.}
}
\description{
Terminates the stack for an application.
}
\section{Accepted Parameters}{
\preformatted{terminate_app(
  appId = "string"
)
}
}

