% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/paws.sms_operations.R
\name{update_replication_job}
\alias{update_replication_job}
\title{The UpdateReplicationJob API is used to change the settings of your existing ReplicationJob created using CreateReplicationJob}
\usage{
update_replication_job(replicationJobId, frequency = NULL,
  nextReplicationRunStartTime = NULL, licenseType = NULL,
  roleName = NULL, description = NULL)
}
\arguments{
\item{replicationJobId}{[required]}

\item{frequency}{}

\item{nextReplicationRunStartTime}{}

\item{licenseType}{}

\item{roleName}{}

\item{description}{}
}
\description{
The UpdateReplicationJob API is used to change the settings of your existing ReplicationJob created using CreateReplicationJob. Calling this API will affect the next scheduled ReplicationRun.
}
\section{Accepted Parameters}{
\preformatted{update_replication_job(
  replicationJobId = "string",
  frequency = 123,
  nextReplicationRunStartTime = as.POSIXct("2015-01-01"),
  licenseType = "AWS"|"BYOL",
  roleName = "string",
  description = "string"
)
}
}

