% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/paws.sms_operations.R
\name{delete_replication_job}
\alias{delete_replication_job}
\title{Deletes the specified replication job}
\usage{
delete_replication_job(replicationJobId)
}
\arguments{
\item{replicationJobId}{[required] The identifier of the replication job.}
}
\description{
Deletes the specified replication job.
}
\details{
After you delete a replication job, there are no further replication runs. AWS deletes the contents of the Amazon S3 bucket used to store AWS SMS artifacts. The AMIs created by the replication runs are not deleted.
}
\section{Accepted Parameters}{
\preformatted{delete_replication_job(
  replicationJobId = "string"
)
}
}

