% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/paws.sms_operations.R
\name{disassociate_connector}
\alias{disassociate_connector}
\title{Disassociates the specified connector from AWS SMS}
\usage{
disassociate_connector(connectorId)
}
\arguments{
\item{connectorId}{[required] The identifier of the connector.}
}
\description{
Disassociates the specified connector from AWS SMS.
}
\details{
After you disassociate a connector, it is no longer available to support replication jobs.
}
\section{Accepted Parameters}{
\preformatted{disassociate_connector(
  connectorId = "string"
)
}
}

