% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/paws.sms_operations.R
\name{disassociate_connector}
\alias{disassociate_connector}
\title{The DisassociateConnector API will disassociate a connector from the Server Migration Service, rendering it unavailable to support replication jobs}
\usage{
disassociate_connector(connectorId)
}
\arguments{
\item{connectorId}{[required]}
}
\description{
The DisassociateConnector API will disassociate a connector from the Server Migration Service, rendering it unavailable to support replication jobs.
}
\section{Accepted Parameters}{
\preformatted{disassociate_connector(
  connectorId = "string"
)
}
}

