% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/paws.sms_operations.R
\name{get_replication_jobs}
\alias{get_replication_jobs}
\title{The GetReplicationJobs API will return all of your ReplicationJobs and their details}
\usage{
get_replication_jobs(replicationJobId = NULL, nextToken = NULL,
  maxResults = NULL)
}
\arguments{
\item{replicationJobId}{}

\item{nextToken}{}

\item{maxResults}{}
}
\description{
The GetReplicationJobs API will return all of your ReplicationJobs and their details. This API returns a paginated list, that may be consecutively called with nextToken to retrieve all ReplicationJobs.
}
\examples{

}
