% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/paws.sms_operations.R
\name{get_replication_runs}
\alias{get_replication_runs}
\title{The GetReplicationRuns API will return all ReplicationRuns for a given ReplicationJob}
\usage{
get_replication_runs(replicationJobId, nextToken = NULL,
  maxResults = NULL)
}
\arguments{
\item{replicationJobId}{[required]}

\item{nextToken}{}

\item{maxResults}{}
}
\description{
The GetReplicationRuns API will return all ReplicationRuns for a given ReplicationJob. This API returns a paginated list, that may be consecutively called with nextToken to retrieve all ReplicationRuns for a ReplicationJob.
}
\section{Accepted Parameters}{
\preformatted{get_replication_runs(
  replicationJobId = "string",
  nextToken = "string",
  maxResults = 123
)
}
}

