% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/paws.sms_operations.R
\name{get_servers}
\alias{get_servers}
\title{The GetServers API returns a list of all servers in your server catalog}
\usage{
get_servers(nextToken = NULL, maxResults = NULL)
}
\arguments{
\item{nextToken}{}

\item{maxResults}{}
}
\description{
The GetServers API returns a list of all servers in your server catalog. For this call to succeed, you must previously have called ImportServerCatalog.
}
\section{Accepted Parameters}{
\preformatted{get_servers(
  nextToken = "string",
  maxResults = 123
)
}
}

