% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/paws.sms_operations.R
\name{import_server_catalog}
\alias{import_server_catalog}
\title{The ImportServerCatalog API is used to gather the complete list of on-premises servers on your premises}
\usage{
import_server_catalog()
}
\description{
The ImportServerCatalog API is used to gather the complete list of on-premises servers on your premises. This API call requires connectors to be installed and monitoring all servers you would like imported. This API call returns immediately, but may take some time to retrieve all of the servers.
}
\section{Accepted Parameters}{
\preformatted{import_server_catalog()
}
}

