% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/paws.sms_package.R
\docType{package}
\name{paws.sms-package}
\alias{paws.sms}
\alias{paws.sms-package}
\title{paws.sms: AWS Server Migration Service}
\description{
Amazon Server Migration Service automates the
process of migrating servers to EC2.
}
\author{
\strong{Maintainer}: David Kretch \email{david.kretch@gmail.com}

Authors:
\itemize{
  \item Adam Banker
}

}
\keyword{internal}
