% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/paws.sms_operations.R
\name{stop_app_replication}
\alias{stop_app_replication}
\title{Stops replicating an application}
\usage{
stop_app_replication(
  appId = "string"
)
}
\arguments{
\item{appId}{ID of the application to stop replicating.}
}
\description{
Stops replicating an application.
}
\examples{

}
